fix: assertEquals delta

+++ src/main/java/Calculator.java
@@ -12,6 +12,6 @@ public class Calculator {
     }
 
     public float divide(int a, int b){
-        return a / b;
+        return (float)a / b;
     }
 }
